add godoc
